aqi-cardio-monitor/backend/config.py
import os
from dotenv import load_dotenv

load_dotenv()


class Config:
    """Application configuration loaded from .env file."""

    # MySQL Database
    DB_HOST = os.getenv('DB_HOST', 'localhost')
    DB_USER = os.getenv('DB_USER', 'root')
    DB_PASSWORD = os.getenv('DB_PASSWORD', '')
    DB_NAME = os.getenv('DB_NAME', 'aqi_cardio_db')

    # OpenWeather API
    OPENWEATHER_API_KEY = os.getenv('OPENWEATHER_API_KEY', '')

    # Flask
    FLASK_DEBUG = os.getenv('FLASK_DEBUG', 'True') == 'True'
    FLASK_PORT = int(os.getenv('FLASK_PORT', 5000))
