aqi-cardio-monitor/backend/requirements.txt
flask
flask-cors
mysql-connector-python
scikit-learn
pandas
numpy
joblib
requests
python-dotenv
werkzeug
